+++ CHANGELOG.md
@@ -33,6 +33,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-2144 Migrate dashboard from blip to Yourloops
 - YLP-2159 Get rid of connectWithTransitionGroup in viz
 - YLP-2162 Migrate CbgDateTrace store from redux to TrendsProvider
+- YLP-2176 Migrate CBG slice actions from redux to TrendsProvider
 
 ### Fixed
 - YLP-2145 Patient cannot see monitoring widgets when he's in a monitoring team and another non-monitoring team
@@ -60,6 +61,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-1625 Display an explicit error when trying to share data with another patient
 - YLP-1680 getPatientData Api method has a wrong parameter
 - YLP-1726 Rework signup with Auth0
+- YLP-1770 Team list is not refreshed after patient accept invitation from notification
 - YLP-1805 Push error to the backend
 - YLP-1868 Add route to invite team member
 - YLP-1873 Fix patient profile update not visible in dashboard